Add WP Loupe core classes and shared trait for enhanced search functionality
