+++ includes/class-wp-loupe-settings.php
